+++ tests/tnt/tnt_test.go
@@ -14,6 +14,9 @@ const (
 
 	envNreplicasetsKey = "NREPLICASETS"
 	envStartPortKey    = "START_PORT"
+
+	defaultTntUser     = "storage"
+	defaultTntPassword = "storage"
 )
 
 func getEnvInt(key string) int {