+++ vshard.go
@@ -59,9 +59,6 @@ type Router struct {
 	// and therefore is global and monotonically growing.
 	refID atomic.Int64
 
-	// worker's count to proceed channel of replicaset's futures
-	nWorkers int32
-
 	cancelDiscovery func()
 }
 
@@ -100,8 +97,6 @@ type Config struct {
 	User             string
 	Password         string
 	PoolOpts         tarantool.Opts
-
-	NWorkers int32 // todo: rename this, cause NWorkers naming looks strange
 }
 
 type BucketStatInfo struct {
@@ -174,13 +169,6 @@ func NewRouter(ctx context.Context, cfg Config) (*Router, error) {
 
 		router.cancelDiscovery = cancelFunc
 	}
-
-	nWorkers := int32(2)
-	if cfg.NWorkers > 0 {
-		nWorkers = cfg.NWorkers
-	}
-
-	router.nWorkers = nWorkers
 
 	return router, err
 }